Separate support for persistence.xml from KeycloakProcessor
