+++ quarkus/deployment/src/main/java/org/keycloak/quarkus/deployment/KeycloakProcessor.java
@@ -51,22 +51,12 @@ import io.quarkus.runtime.configuration.ConfigurationException;
 import io.quarkus.vertx.http.deployment.HttpRootPathBuildItem;
 import io.quarkus.vertx.http.deployment.NonApplicationRootPathBuildItem;
 import io.quarkus.vertx.http.deployment.RouteBuildItem;
-import jakarta.persistence.Entity;
-import jakarta.persistence.PersistenceUnitTransactionType;
 import org.eclipse.microprofile.config.spi.ConfigSource;
-import org.hibernate.cfg.JdbcSettings;
 import org.eclipse.microprofile.health.Readiness;
-import org.hibernate.cfg.AvailableSettings;
-import org.hibernate.jpa.boot.internal.ParsedPersistenceXmlDescriptor;
-import org.hibernate.jpa.boot.spi.PersistenceUnitDescriptor;
-import org.hibernate.jpa.boot.spi.PersistenceXmlParser;
 import org.infinispan.protostream.SerializationContextInitializer;
-import org.jboss.jandex.AnnotationInstance;
-import org.jboss.jandex.AnnotationTarget;
 import org.jboss.jandex.AnnotationTransformation;
 import org.jboss.jandex.ClassInfo;
 import org.jboss.jandex.DotName;
-import org.jboss.jandex.IndexView;
 import org.jboss.jandex.MethodInfo;
 import org.jboss.logging.Logger;
 import org.jboss.resteasy.reactive.server.model.HandlerChainCustomizer;
@@ -80,7 +70,6 @@ import org.keycloak.common.Profile;
 import org.keycloak.common.crypto.FipsMode;
 import org.keycloak.common.util.MultiSiteUtils;
 import org.keycloak.common.util.StreamUtil;
-import org.keycloak.config.DatabaseOptions;
 import org.keycloak.config.HealthOptions;
 import org.keycloak.config.HttpOptions;
 import org.keycloak.config.LoggingOptions;
@@ -90,7 +79,6 @@ import org.keycloak.config.SecurityOptions;
 import org.keycloak.config.TracingOptions;
 import org.keycloak.config.TransactionOptions;
 import org.keycloak.connections.jpa.DefaultJpaConnectionProviderFactory;
-import org.keycloak.connections.jpa.JpaConnectionProvider;
 import org.keycloak.connections.jpa.JpaConnectionSpi;
 import org.keycloak.connections.jpa.updater.liquibase.LiquibaseJpaUpdaterProviderFactory;
 import org.keycloak.connections.jpa.updater.liquibase.conn.DefaultLiquibaseConnectionProvider;
@@ -116,7 +104,6 @@ import org.keycloak.quarkus.runtime.integration.resteasy.KeycloakHandlerChainCus
 import org.keycloak.quarkus.runtime.integration.resteasy.KeycloakTracingCustomizer;
 import org.keycloak.quarkus.runtime.logging.ClearMappedDiagnosticContextFilter;
 import org.keycloak.quarkus.runtime.services.health.KeycloakReadyHealthCheck;
-import org.keycloak.quarkus.runtime.storage.database.jpa.NamedJpaConnectionProviderFactory;
 import org.keycloak.quarkus.runtime.themes.FlatClasspathThemeResourceProviderFactory;
 import org.keycloak.representations.provider.ScriptProviderDescriptor;
 import org.keycloak.representations.provider.ScriptProviderMetadata;
@@ -144,8 +131,6 @@ import java.net.URL;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Path;
 import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Collections;
 import java.util.Enumeration;
 import java.util.HashMap;
 import java.util.List;
@@ -156,21 +141,15 @@ import java.util.Optional;
 import java.util.Properties;
 import java.util.ServiceLoader;
 import java.util.Set;
-import java.util.function.BiConsumer;
 import java.util.function.Function;
-import java.util.function.Predicate;
 import java.util.jar.JarEntry;
 import java.util.jar.JarFile;
 import java.util.logging.Handler;
 
-import static org.keycloak.connections.jpa.util.JpaUtils.loadSpecificNamedQueries;
 import static org.keycloak.quarkus.runtime.Environment.getCurrentOrCreateFeatureProfile;
 import static org.keycloak.quarkus.runtime.Providers.getProviderManager;
-import static org.keycloak.quarkus.runtime.configuration.Configuration.getOptionalBooleanKcValue;
-import static org.keycloak.quarkus.runtime.configuration.Configuration.getOptionalKcValue;
 import static org.keycloak.quarkus.runtime.configuration.Configuration.getOptionalValue;
 import static org.keycloak.quarkus.runtime.configuration.MicroProfileConfigProvider.NS_KEYCLOAK_PREFIX;
-import static org.keycloak.quarkus.runtime.storage.database.jpa.QuarkusJpaConnectionProviderFactory.QUERY_PROPERTY_PREFIX;
 import static org.keycloak.representations.provider.ScriptProviderDescriptor.AUTHENTICATORS;
 import static org.keycloak.representations.provider.ScriptProviderDescriptor.MAPPERS;
 import static org.keycloak.representations.provider.ScriptProviderDescriptor.POLICIES;
@@ -370,47 +349,6 @@ class KeycloakProcessor {
     }
 
     /**
-     * Get datasource name obtained from the persistence.xml file based on this order:
-     * <ol>
-     *      <li> return {@link JdbcSettings#JAKARTA_JTA_DATASOURCE} if specified
-     *      <li> return {@link AvailableSettings#DATASOURCE} property if specified
-     *      <li> return persistence unit name
-     * </ol>
-     * Can be removed after removing support for persistence.xml files
-     */
-    static String getDatasourceNameFromPersistenceXml(PersistenceUnitDescriptor descriptor) {
-        if (descriptor == null) {
-            throw new IllegalStateException("Descriptor cannot be null");
-        }
-        final BiConsumer<String, String> infoAboutUsedSourceForDsName = (source, name) -> logger.debugf(
-                "Datasource name '%s' is obtained from the '%s' configuration property in persistence.xml file. " +
-                        "Use '%s' name for datasource options like 'db-kind-%s'.", name, source, name, name);
-
-        String persistenceUnitName = descriptor.getName();
-        Properties properties = descriptor.getProperties();
-
-        // 1. return Jakarta properties
-        var jakartaProperty = properties.getProperty(JdbcSettings.JAKARTA_JTA_DATASOURCE);
-        if (jakartaProperty != null) {
-            infoAboutUsedSourceForDsName.accept(JdbcSettings.JAKARTA_JTA_DATASOURCE, jakartaProperty);
-            return jakartaProperty;
-        }
-
-        // 2. return deprecated Hibernate property
-        var deprecatedHibernateProperty = properties.getProperty(AvailableSettings.DATASOURCE);
-        if (deprecatedHibernateProperty != null) {
-            logger.warnf("Property '%s' is deprecated for some time and you should rather use '%s' property for datasource name in persistence.xml file.",
-                    AvailableSettings.DATASOURCE, JdbcSettings.JAKARTA_JTA_DATASOURCE);
-            infoAboutUsedSourceForDsName.accept(AvailableSettings.DATASOURCE, deprecatedHibernateProperty);
-            return deprecatedHibernateProperty;
-        }
-
-        // 3. return persistence unit name
-        infoAboutUsedSourceForDsName.accept("Persistence unit name", persistenceUnitName);
-        return persistenceUnitName;
-    }
-
-    /**
      * <p>Configures the persistence unit for Quarkus.
      *
      * <p>The {@code hibernate-orm} extension expects that the dialect is statically
@@ -430,122 +368,16 @@ class KeycloakProcessor {
             CombinedIndexBuildItem indexBuildItem,
             BuildProducer<HibernateOrmIntegrationRuntimeConfiguredBuildItem> runtimeConfigured,
             KeycloakRecorder recorder) {
-        ParsedPersistenceXmlDescriptor defaultUnitDescriptor = null;
-        List<String> userManagedEntities = new ArrayList<>();
-
-        for (PersistenceXmlDescriptorBuildItem item : descriptors) {
-            ParsedPersistenceXmlDescriptor descriptor = (ParsedPersistenceXmlDescriptor) item.getDescriptor();
-
-            if ("keycloak-default".equals(descriptor.getName())) {
-                defaultUnitDescriptor = descriptor;
-                configureDefaultPersistenceUnitProperties(defaultUnitDescriptor, config, getDefaultDataSource(jdbcDataSources));
-                runtimeConfigured.produce(new HibernateOrmIntegrationRuntimeConfiguredBuildItem("keycloak", defaultUnitDescriptor.getName())
-                        .setInitListener(recorder.createDefaultUnitListener()));
-            } else {
-                String datasourceName = getDatasourceNameFromPersistenceXml(descriptor);
-                configurePersistenceUnitProperties(datasourceName, descriptor);
-                // register a listener for customizing the unit configuration at runtime
-                runtimeConfigured.produce(new HibernateOrmIntegrationRuntimeConfiguredBuildItem("keycloak", descriptor.getName())
-                        .setInitListener(recorder.createUserDefinedUnitListener(datasourceName)));
-                userManagedEntities.addAll(descriptor.getManagedClassNames());
-            }
-        }
-
-        if (defaultUnitDescriptor == null) {
-            throw new RuntimeException("No default persistence unit found.");
-        }
-
-        configureDefaultPersistenceUnitEntities(defaultUnitDescriptor, indexBuildItem, userManagedEntities);
+        PersistenceXmlSupportUtils.configurePersistenceUnits(config, descriptors, jdbcDataSources, indexBuildItem, runtimeConfigured, recorder);
     }
 
     @BuildStep
     @Consume(CheckJdbcBuildStep.class)
     @Consume(CheckMultipleDatasourcesBuildStep.class)
     void produceDefaultPersistenceUnit(BuildProducer<PersistenceXmlDescriptorBuildItem> producer) {
-        PersistenceXmlParser parser = PersistenceXmlParser.create();
-        PersistenceUnitDescriptor descriptor = parser.parse(Collections.singletonList(parser.getClassLoaderService().locateResource("default-persistence.xml")))
-                .values()
-                .stream()
-                .findAny()
+        var defaultPersistenceUnit = PersistenceXmlSupportUtils.getDefaultPersistenceUnit()
                 .orElseThrow(() -> new NoSuchElementException("Cannot find the file 'default-persistence.xml'"));
-
-        producer.produce(new PersistenceXmlDescriptorBuildItem(descriptor));
-    }
-
-    static void configurePersistenceUnitProperties(String datasourceName, ParsedPersistenceXmlDescriptor descriptor) {
-        Properties unitProperties = descriptor.getProperties();
-        var isResourceLocalSpecified = PersistenceUnitTransactionType.RESOURCE_LOCAL.equals(descriptor.getPersistenceUnitTransactionType()) ||
-                Optional.ofNullable(unitProperties.getProperty(AvailableSettings.JAKARTA_TRANSACTION_TYPE))
-                        .map(f -> f.equalsIgnoreCase(PersistenceUnitTransactionType.RESOURCE_LOCAL.name()))
-                        .orElse(false);
-        if (isResourceLocalSpecified) {
-            throw new IllegalArgumentException("You need to use '%s' transaction type in your persistence.xml file."
-                    .formatted(PersistenceUnitTransactionType.JTA.name()));
-        }
-
-        unitProperties.setProperty(AvailableSettings.JAKARTA_TRANSACTION_TYPE, PersistenceUnitTransactionType.JTA.name());
-        descriptor.setTransactionType(PersistenceUnitTransactionType.JTA);
-
-        // set datasource name
-        unitProperties.setProperty(JdbcSettings.JAKARTA_JTA_DATASOURCE,datasourceName);
-        unitProperties.setProperty(AvailableSettings.DATASOURCE, datasourceName); // for backward compatibility
-
-        DatabaseOptions.getNamedKey(DatabaseOptions.DB_SQL_JPA_DEBUG, datasourceName)
-                .filter(Configuration::isKcPropertyTrue)
-                .ifPresent(f -> unitProperties.put(AvailableSettings.USE_SQL_COMMENTS, "true"));
-
-        DatabaseOptions.getNamedKey(DatabaseOptions.DB_SQL_LOG_SLOW_QUERIES, datasourceName)
-                .flatMap(Configuration::getOptionalKcValue)
-                .ifPresent(threshold -> unitProperties.put(AvailableSettings.LOG_SLOW_QUERY, threshold));
-    }
-
-    private void configureDefaultPersistenceUnitProperties(ParsedPersistenceXmlDescriptor descriptor, HibernateOrmConfig config,
-            JdbcDataSourceBuildItem defaultDataSource) {
-        if (defaultDataSource == null || !defaultDataSource.isDefault()) {
-            throw new RuntimeException("The server datasource must be the default datasource.");
-        }
-
-        Properties unitProperties = descriptor.getProperties();
-
-        final Optional<String> dialect = getOptionalKcValue(DatabaseOptions.DB_DIALECT.getKey());
-        dialect.ifPresent(d -> unitProperties.setProperty(AvailableSettings.DIALECT, d));
-
-        final Optional<String> defaultSchema = getOptionalKcValue(DatabaseOptions.DB_SCHEMA.getKey());
-        defaultSchema.ifPresent(ds -> unitProperties.setProperty(AvailableSettings.DEFAULT_SCHEMA, ds));
-
-        unitProperties.setProperty(AvailableSettings.JAKARTA_TRANSACTION_TYPE, PersistenceUnitTransactionType.JTA.name());
-        descriptor.setTransactionType(PersistenceUnitTransactionType.JTA);
-
-        unitProperties.setProperty(AvailableSettings.QUERY_STARTUP_CHECKING, Boolean.FALSE.toString());
-
-        String dbKind = defaultDataSource.getDbKind();
-
-        for (Entry<Object, Object> query : loadSpecificNamedQueries(dbKind.toLowerCase()).entrySet()) {
-            unitProperties.setProperty(QUERY_PROPERTY_PREFIX + query.getKey(), query.getValue().toString());
-        }
-
-        if (getOptionalBooleanKcValue(DatabaseOptions.DB_SQL_JPA_DEBUG.getKey()).orElse(false)) {
-            unitProperties.put(AvailableSettings.USE_SQL_COMMENTS, "true");
-        }
-
-        getOptionalKcValue(DatabaseOptions.DB_SQL_LOG_SLOW_QUERIES.getKey())
-                .ifPresent(v -> unitProperties.put(AvailableSettings.LOG_SLOW_QUERY, v));
-    }
-
-    private void configureDefaultPersistenceUnitEntities(ParsedPersistenceXmlDescriptor descriptor, CombinedIndexBuildItem indexBuildItem,
-            List<String> userManagedEntities) {
-        IndexView index = indexBuildItem.getIndex();
-        Collection<AnnotationInstance> annotations = index.getAnnotations(DotName.createSimple(Entity.class.getName()));
-
-        for (AnnotationInstance annotation : annotations) {
-            AnnotationTarget target = annotation.target();
-            String targetName = target.asClass().name().toString();
-
-            if (!userManagedEntities.contains(targetName)
-                    && (!targetName.startsWith("org.keycloak") || targetName.startsWith("org.keycloak.testsuite"))) {
-                descriptor.addClasses(targetName);
-            }
-        }
+        producer.produce(new PersistenceXmlDescriptorBuildItem(defaultPersistenceUnit));
     }
 
     /**
@@ -580,7 +412,7 @@ class KeycloakProcessor {
             }
 
             if (spi instanceof JpaConnectionSpi) {
-                configureUserDefinedPersistenceUnits(descriptors, factories, preConfiguredProviders, spi);
+                PersistenceXmlSupportUtils.configureUserDefinedPersistenceUnits(descriptors, factories, preConfiguredProviders, spi);
             }
 
             if (spi instanceof ThemeResourceSpi) {
@@ -620,22 +452,7 @@ class KeycloakProcessor {
         }
     }
 
-    private void configureUserDefinedPersistenceUnits(List<PersistenceXmlDescriptorBuildItem> descriptors,
-            Map<Spi, Map<Class<? extends Provider>, Map<String, Class<? extends ProviderFactory>>>> factories,
-            Map<String, ProviderFactory> preConfiguredProviders, Spi spi) {
-        descriptors.stream()
-                .map(PersistenceXmlDescriptorBuildItem::getDescriptor)
-                .map(PersistenceUnitDescriptor::getName)
-                .filter(Predicate.not("keycloak-default"::equals))
-                .forEach((String unitName) -> {
-                    NamedJpaConnectionProviderFactory factory = new NamedJpaConnectionProviderFactory();
 
-                    factory.setUnitName(unitName);
-
-                    factories.get(spi).get(JpaConnectionProvider.class).put(unitName, NamedJpaConnectionProviderFactory.class);
-                    preConfiguredProviders.put(unitName, factory);
-                });
-    }
 
     /**
      * Register the custom {@link ConfigSource} implementations.
