+++ quarkus/deployment/src/test/java/org/keycloak/quarkus/deployment/PersistenceXmlDatasourcesTest.java
@@ -24,8 +24,8 @@ import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.keycloak.quarkus.deployment.KeycloakProcessor.configurePersistenceUnitProperties;
-import static org.keycloak.quarkus.deployment.KeycloakProcessor.getDatasourceNameFromPersistenceXml;
+import static org.keycloak.quarkus.deployment.PersistenceXmlSupportUtils.configurePersistenceUnitProperties;
+import static org.keycloak.quarkus.deployment.PersistenceXmlSupportUtils.getDatasourceNameFromPersistenceXml;
 import static org.wildfly.common.Assert.assertNotNull;
 
 public class PersistenceXmlDatasourcesTest {
